refactor: Major refactor of packer scripts.
Now broken into three sections.  "Production" folders, "local" folders
for debugging only, and "test" folders for automated test runs.  Lots of
duplication among the packer files, but with packer that is not
avoidable.  Will continue to work on reducing duplication in bash
scripts.  May also convert to python instead of bash as a build system.